+++ src/rs_driver/driver/decoder/decoder_RS80.hpp
@@ -112,8 +112,8 @@ typedef struct
   uint8_t reserved_2[5];
   RSDiagno diagno;
   uint8_t gprmc[86];
-  RSCorAngle ver_angle_cali[128];
-  RSCorAngle hori_angle_cali[128];
+  RSCalibrationAngle ver_angle_cali[128];
+  RSCalibrationAngle hori_angle_cali[128];
   uint8_t reserved_3[10];
   uint16_t tail;
 } RS80DifopPkt;
@@ -141,7 +141,9 @@ private:
 template <typename T_Point>
 DecoderRS80<T_Point>::DecoderRS80(const RSDecoderParam& param) : DecoderBase<T_Point>(param)
 {
-  this->angle_file_index_ = 128;
+  this->angle_file_row_num_ = 128;
+  this->vert_angle_list_.resize(this->angle_file_row_num_);
+  this->hori_angle_list_.resize(this->angle_file_row_num_);
   if (this->param_.max_distance > 230.0f)
   {
     this->param_.max_distance = 230.0f;
@@ -209,8 +211,6 @@ RSDecoderResult DecoderRS80<T_Point>::decodeMsopPkt(const uint8_t* pkt, std::vec
       break;
     }
     int cur_azi = RS_SWAP_SHORT(mpkt_ptr->blocks[blk_idx].azimuth);
-    float azi_diff_theoretical =
-        (360 / RS80_BLOCKS_PER_PKT) * 100 / (float)this->pkts_per_frame_;  ///< ((rpm/60)*360)/pkts_rate/blocks_per_pkt
     if (this->echo_mode_ == ECHO_DUAL)
     {
       if (blk_idx % 2 == 0)
@@ -244,7 +244,7 @@ RSDecoderResult DecoderRS80<T_Point>::decodeMsopPkt(const uint8_t* pkt, std::vec
                                              (block_timestamp + this->time_duration_between_blocks_);
       }
     }
-    azi_diff = (azi_diff > 100) ? azi_diff_theoretical : azi_diff;
+    azi_diff = (azi_diff > 100) ? this->azi_diff_between_block_theoretical_ : azi_diff;
     for (size_t channel_idx = 0; channel_idx < RS80_CHANNELS_PER_BLOCK; channel_idx++)
     {
       int dsr_temp = (channel_idx / 4) % 16;
@@ -316,6 +316,8 @@ RSDecoderResult DecoderRS80<T_Point>::decodeDifopPkt(const uint8_t* pkt)
   {
     this->pkts_per_frame_ = ceil(RS80_PKT_RATE * 60 / this->rpm_);
   }
+  this->azi_diff_between_block_theoretical_ =
+      (RS_ONE_ROUND / RS80_BLOCKS_PER_PKT) / (float)this->pkts_per_frame_;  ///< ((rpm/60)*360)/pkts_rate/blocks_per_pkt
   if (!this->difop_flag_)
   {
     bool angle_flag = true;
@@ -330,7 +332,7 @@ RSDecoderResult DecoderRS80<T_Point>::decodeDifopPkt(const uint8_t* pkt)
     if (angle_flag)
     {
       int lsb, mid, msb, neg = 1;
-      for (size_t i = 0; i < this->angle_file_index_; i++)
+      for (size_t i = 0; i < this->angle_file_row_num_; i++)
       {
         // calculation of vertical angle
         lsb = dpkt_ptr->ver_angle_cali[i].sign;