+++ src/rs_driver/driver/decoder/decoder_factory.hpp
@@ -35,117 +35,62 @@ class DecoderFactory
 {
 public:
   inline static std::shared_ptr<DecoderBase<T_Point>> createDecoder(const LidarType& param_lidar_type,
-                                                                   const RSDecoderParam& param,
-                                                                   const PacketMsg& msop_pkt_msg,
-                                                                   const std::shared_ptr<Input>& input_ptr)
+                                                                    const RSDriverParam& param,
+                                                                    const PacketMsg& msop_pkt_msg,
+                                                                    const std::shared_ptr<Input>& input_ptr)
   {
-    if (param_lidar_type == LidarType::RSAUTO)
-    {
-      RSMsopHeader* header_ptr = (RSMsopHeader*)msop_pkt_msg.packet.data();
-      input_ptr->setLidarType((LidarType)header_ptr->lidar_type);
-      switch (header_ptr->lidar_type)
-      {
-        case LidarType::RS16:
-          return std::make_shared<DecoderRS16<T_Point>>(param);
-          break;
-        case LidarType::RS32:
-          return std::make_shared<DecoderRS32<T_Point>>(param);
-          break;
-        case LidarType::RSBP:
-          return std::make_shared<DecoderRSBP<T_Point>>(param);
-          break;
-        case LidarType::RS128:
-          return std::make_shared<DecoderRS128<T_Point>>(param);
-          break;
-        case LidarType::RS80:
-          return std::make_shared<DecoderRS80<T_Point>>(param);
-          break;
-        default:
-          ERROR << "Wrong LiDAR Type. Please check your LiDAR Version! " << REND;
-          exit(-1);
-      }
-    }
-    else
-    {
-      input_ptr->setLidarType(param_lidar_type);
-      switch (param_lidar_type)
-      {
-        case LidarType::RS16:
-          return std::make_shared<DecoderRS16<T_Point>>(param);
-          break;
-        case LidarType::RS32:
-          return std::make_shared<DecoderRS32<T_Point>>(param);
-          break;
-        case LidarType::RSBP:
-          return std::make_shared<DecoderRSBP<T_Point>>(param);
-          break;
-        case LidarType::RS128:
-          return std::make_shared<DecoderRS128<T_Point>>(param);
-          break;
-        case LidarType::RS80:
-          return std::make_shared<DecoderRS80<T_Point>>(param);
-          break;
-        default:
-          ERROR << "Wrong LiDAR Type. Please check your LiDAR Version! " << REND;
-          exit(-1);
-      }
-    }
+    LidarType lidar_type = getLidarType(param_lidar_type, msop_pkt_msg);
+    input_ptr->setLidarType(lidar_type);
+    return switchLidar(lidar_type, param);
   }
   inline static std::shared_ptr<DecoderBase<T_Point>> createDecoder(const LidarType& param_lidar_type,
-                                                                   const RSDecoderParam& param,
-                                                                   const PacketMsg& msop_pkt_msg)
+                                                                    const RSDriverParam& param,
+                                                                    const PacketMsg& msop_pkt_msg)
+  {
+    LidarType lidar_type = getLidarType(param_lidar_type, msop_pkt_msg);
+    return switchLidar(lidar_type, param);
+  }
+
+  inline static LidarType getLidarType(const LidarType& param_lidar_type, const PacketMsg& msop_pkt_msg)
   {
     if (param_lidar_type == LidarType::RSAUTO)
     {
       RSMsopHeader* header_ptr = (RSMsopHeader*)msop_pkt_msg.packet.data();
-      switch (header_ptr->lidar_type)
-      {
-        case LidarType::RS16:
-          return std::make_shared<DecoderRS16<T_Point>>(param);
-          break;
-        case LidarType::RS32:
-          return std::make_shared<DecoderRS32<T_Point>>(param);
-          break;
-        case LidarType::RSBP:
-          return std::make_shared<DecoderRSBP<T_Point>>(param);
-          break;
-        case LidarType::RS128:
-          return std::make_shared<DecoderRS128<T_Point>>(param);
-          break;
-        case LidarType::RS80:
-          return std::make_shared<DecoderRS80<T_Point>> (param);
-          break;
-        default:
-          ERROR << "Wrong LiDAR Type. Please check your LiDAR Version! " << REND;
-          exit(-1);
-      }
+      return (LidarType)header_ptr->lidar_type;
     }
     else
     {
-      switch (param_lidar_type)
-      {
-        case LidarType::RS16:
-          return std::make_shared<DecoderRS16<T_Point>>(param);
-          break;
-        case LidarType::RS32:
-          return std::make_shared<DecoderRS32<T_Point>>(param);
-          break;
-        case LidarType::RSBP:
-          return std::make_shared<DecoderRSBP<T_Point>>(param);
-          break;
-        case LidarType::RS128:
-          return std::make_shared<DecoderRS128<T_Point>>(param);
-          break;
-        case LidarType::RS80:
-          return std::make_shared<DecoderRS80<T_Point>>(param);
-          break;
-        default:
-          ERROR << "Wrong LiDAR Type. Please check your LiDAR Version! " << REND;
-          exit(-1);
-      }
+      return param_lidar_type;
     }
   }
+  inline static std::shared_ptr<DecoderBase<T_Point>> switchLidar(const LidarType& lidar_type,
+                                                                  const RSDriverParam& param)
+  {
+    std::shared_ptr<DecoderBase<T_Point>> ret_ptr;
+    switch (lidar_type)
+    {
+      case LidarType::RS16:
+        ret_ptr = std::make_shared<DecoderRS16<T_Point>>(param.decoder_param);
+        break;
+      case LidarType::RS32:
+        ret_ptr = std::make_shared<DecoderRS32<T_Point>>(param.decoder_param);
+        break;
+      case LidarType::RSBP:
+        ret_ptr = std::make_shared<DecoderRSBP<T_Point>>(param.decoder_param);
+        break;
+      case LidarType::RS128:
+        ret_ptr = std::make_shared<DecoderRS128<T_Point>>(param.decoder_param);
+        break;
+      case LidarType::RS80:
+        ret_ptr = std::make_shared<DecoderRS80<T_Point>>(param.decoder_param);
+        break;
+      default:
+        ERROR << "Wrong LiDAR Type. Please check your LiDAR Version! " << REND;
+        exit(-1);
+    }
+    ret_ptr->loadCalibrationFile(param.angle_path);
+    return ret_ptr;
+  }
 };
-
 }  // namespace lidar
 }  // namespace robosense